CC-9565: Add support for casting stringified UUIDs to uuid upon insert/update/upsert in PostgreSQL
Modify the methods on the `DatabaseDialect` interface used to supply the insert, update, and upsert expressions and create the prepared statements to use the table definition (and the corresponding the column definitions). The `GenericDatabaseDialect` class does not use these definitions, but the `PostgreSqlDatabaseDialect` class is modified to add the cast expression (e.g., `::uuid`) to the variable placeholder (e.g, `?`) if the column type matches one of the database types that requires casting, which in this case is either `uuid`, `json`, or `jsonb`.

Added an integration test to verify that string UUID and JSON literals can be inserted into a table with a `uuid` column and a `json` column.